+++ packages/web3-strategy/src/types/strategy.ts
@@ -1,6 +1,5 @@
 import { AccountAddress, ChainId } from '@injectivelabs/ts-types'
 import Web3 from 'web3'
-import { provider } from 'web3-core'
 
 export type onAccountChangeCallback = (account: AccountAddress) => void
 export type onChainIdChangeCallback = () => void
@@ -14,44 +13,14 @@ export interface ConcreteStrategyOptions {
   privateKey?: string
   baseDerivationPath?: string
   derivationPathType?: LedgerDerivationPathType
-  rpcUrls?: Record<ChainId, string>
-  wsRpcUrls?: Record<ChainId, string>
-  pollingInterval: number
-  blockTracker?: boolean
+  rpcUrl: string
+  wsRpcUrl: string
 }
 
 export interface ConcreteWeb3Strategy {
-  web3ForChainId: Record<ChainId, Web3>
-
-  web3WsForChainId: Record<ChainId, Web3>
-
-  getWeb3ProviderEngineForRpc({
-    rpcUrl,
-    pollingInterval,
-    blockTracker,
-  }: {
-    rpcUrl: string
-    pollingInterval: number
-    blockTracker?: any
-  }): provider
-
-  getWeb3WsProviderEngineForRpc({
-    wsRpcUrl,
-    pollingInterval,
-    blockTracker,
-  }: {
-    wsRpcUrl: string
-    pollingInterval: number
-    blockTracker?: any
-  }): provider
+  web3: Web3
 
   getWeb3(): Web3
-
-  getWeb3ForChainId(chainId: ChainId): Web3
-
-  getWeb3Ws(): Web3
-
-  getWeb3WsForChainId(chainId: ChainId): Web3
 
   isMetamask(): boolean