+++ packages/web3-strategy/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.89](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/web3-strategy@0.0.88...@injectivelabs/web3-strategy@0.0.89) (2022-02-09)
+
+**Note:** Version bump only for package @injectivelabs/web3-strategy
+
+
+
+
+
 ## [0.0.88](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/web3-strategy@0.0.87...@injectivelabs/web3-strategy@0.0.88) (2022-02-08)